removed a few harmless warnings

+++ src/meshlabplugins/filter_isoparametrization/iso_parametrization.h
@@ -1722,7 +1722,7 @@ public:
 	template <class MeshType>
 	void CopyParametrization(MeshType	* _param_mesh)
 	{
-		for (int i=0;i<_param_mesh->vert.size();i++)
+    for (size_t i=0;i<_param_mesh->vert.size();i++)
 		{
 			_param_mesh->vert[i].T().P()=ParaMesh()->vert[i].T().P();
 			_param_mesh->vert[i].T().N()=ParaMesh()->vert[i].T().N();
